Compile on float types

+++ src/matplotlibcpp.cpp
@@ -386,7 +386,7 @@ bool plot(
     const Eigen::Ref<const VectorX>& y,
     const std::string& format)
 {
-  Eigen::Matrix<double, Eigen::Dynamic, 1> x(y.size());
+  Eigen::Matrix<FloatType, Eigen::Dynamic, 1> x(y.size());
   for (int i=0; i < x.size(); ++i)
   {
     x(i) = i;
